requirements.txt
aiohttp
aiosqlite
discord.py==2.6.4
python-dotenv
supabase
